Add support for SimpleForm.
The type needs to be called :pwdcalc_simple to avoid a nameclash when using both Formtastic and SimpleForm.
