+++ lib/pwdcalc.rb
@@ -1,4 +1,5 @@
 require "pwdcalc/helpers/hints"
+require 'pwdcalc/password_strength'
 
 module Pwdcalc
   # Required for jquery.YAPSM.min.js and jquery.pwdcalc.js to be discoverable in the asset pipeline
@@ -8,10 +9,13 @@ module Pwdcalc
     end
 
     initializer "pwdcalc.initialize" do
-
       # Add custom input to Formtastic if it's included in Gemfile
       if defined?(Formtastic)
         require 'pwdcalc/formtastic'
+      end
+
+      if defined?(SimpleForm)
+        require 'pwdcalc/simple_form'
       end
 
       ActionView::Base.send :include, Pwdcalc::Helpers::Hints
