+++ lib/pwdcalc/formtastic.rb
@@ -4,6 +4,7 @@ module Formtastic
 
       include Base
       include Base::Stringish
+      include Pwdcalc::PasswordStrength
 
       def to_html
         input_wrapping do
@@ -12,18 +13,6 @@ module Formtastic
           password_strength_score <<
           password_strength_meter
         end
-      end
-
-      def password_strength_score
-        template.content_tag(:div, :class => "pwdcalc-score") do
-          template.content_tag(:span) do
-            template.content_tag(:b, "")
-          end
-        end
-      end
-
-      def password_strength_meter
-        template.content_tag(:p, "", :'data-field' => input_html_options[:id], :class => "password-strength-meter pwdcalc")
       end
     end
   end
